Moves widgets to widgets module
