+++ lib/tictactoe/gui/qtgui/menu_gui.rb
@@ -1,3 +1,5 @@
+require 'tictactoe/gui/qtgui/widgets/factory'
+
 module Tictactoe
   module Gui
     module QtGui
@@ -11,7 +13,7 @@ module Tictactoe
 
         private
         def init  
-          @widget_factory = QtGui::WidgetFactory.new()
+          @widget_factory = QtGui::Widgets::Factory.new()
           @window = @widget_factory.new_window(240, 150)
           @game_options = @widget_factory.new_game_options(@on_configured)
           @widget_factory.layout(@window, [@game_options])
