+++ spec/tictactoe/gui/game_window_spec.rb
@@ -1,6 +1,5 @@
 require 'spec_helper'
 require 'tictactoe/gui/game_window'
-require 'tictactoe/gui/qtgui/factory'
 require 'tictactoe/gui/qtgui/game_gui'
 
 RSpec.describe Tictactoe::Gui::GameWindow, :integration => true, :gui => true do
